HealthMonitor: Add MON_NETSPLIT warning
Problem:

Currently, We don't have a way to warn users
when Netsplit is happening between two or more CRUSH buckets.

Solution:

We can implement something that could utilize the Connection Report
of the Monitors when we are using the Connectivity Election Strategy
to determine if there is a Netsplit in the cluster. If there is a Monitor
that can reach everyone but it sees that in the peer_scores there
are peers that cannot reach one another.
We can report which pair of buckets are experiencing Netsplit.

Fixes: https://tracker.ceph.com/issues/67371

Signed-off-by: Kamoltat Sirivadhna <ksirivad@redhat.com>