[Stretch Mode] mon: restrict changing mon election strategy post stretch mode
Problem:

It is possible to modify the monitor election
strategy once stretch mode is enabled.

Solution:

Restrict the ability for Ceph to change election strategy
while using stretch mode.

Also, notify the user if they attempt to change to the current election
strategy. This change prevents unnecessary resource usage by informing
the user when they try to switch to an election strategy that is already in use.
This avoids redundant updates to the monmap.

Fixes: https://tracker.ceph.com/issues/67801

Signed-off-by: Kamoltat Sirivadhna <ksirivad@redhat.com>